Improving Network synch 56

+++ CppScripts/BBBhw/GPIO.cpp
@@ -852,7 +852,7 @@ for (iIterDump=0; iIterDump<NumQuBitsPerRun; iIterDump++){
 }
 ////////////////////////////////////////////
 // Checks of proper values handling
-unsigned long long int CheckValueAux=static_cast<unsigned long long int>((static_cast<long long int>(SynchTrigPeriod/2.0)+static_cast<long long int>(TimeTaggsStored[0]))%static_cast<long long int>(SynchTrigPeriod)-static_cast<long long int>(SynchTrigPeriod/2.0));
+long long int CheckValueAux=(static_cast<long long int>(SynchTrigPeriod/2.0)+static_cast<long long int>(TimeTaggsStored[0]))%static_cast<long long int>(SynchTrigPeriod)-static_cast<long long int>(SynchTrigPeriod/2.0);
 cout << "GPIO::DDRdumpdata::CheckValueAux: "<< CheckValueAux << endl;
 cout << "GPIO::DDRdumpdata::SynchTrigPeriod: " << SynchTrigPeriod << endl;
 cout << "GPIO::DDRdumpdata::NumQuBitsPerRun: " << NumQuBitsPerRun << endl;
@@ -1078,7 +1078,7 @@ for (int i=0;i<TotalCurrentNumRecords;i++){
 
 //////////////////////////////////////////
 // Checks of proper values handling
-unsigned long long int CheckValueAux=static_cast<unsigned long long int>((static_cast<long long int>(SynchTrigPeriod/2.0)+static_cast<long long int>(TimeTaggsStored[0]))%static_cast<long long int>(SynchTrigPeriod)-static_cast<long long int>(SynchTrigPeriod/2.0));
+long long int CheckValueAux=(static_cast<long long int>(SynchTrigPeriod/2.0)+static_cast<long long int>(TimeTaggsStored[0]))%static_cast<long long int>(SynchTrigPeriod)-static_cast<long long int>(SynchTrigPeriod/2.0);
 cout << "GPIO::PRUdetCorrRelFreq::CheckValueAux: "<< CheckValueAux << endl;
 ////////////////////////////////////////
 return 0; // All ok
